refactor: 채팅방 화면 타입 정의 파일 통합 및 적용
